requirements.txt
livekit-agents[deepgram,openai,cartesia,silero,turn-detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
livekit-plugins-openai
livekit-plugins-silero
livekit-api
pandas
signalrcore